websocket-ssl-acceptor: provide request/reply-class default-initargs
acceptor-dispatch-request in hunchensocket is specialized on both
acceptor and requst, where the latter must be websocket-request.

websocket-ssl-acceptor does inherit from websocket-acceptor, but it
doesn't set default-initargs for request and reply classes, what leads
to wrong dispatch, where default acceptor-dispatch-request is chosen
instead of websocket one (because it has second argument specialized
on T), what leads to 404 on all ssl connections.