Replace ironclad dependency with sha1
Ironclad is a meaty dependency. This system takes the longest time on my
computer to compile my web application, and I've read comments online
that it adds ~19MB RAM usage in a lisp image. Hunchensocket is
currently using ironclad to produce SHA1 digests. A single-purpose
library can fill this need instead.

This change switches the dependency from ironclad:digest-sequence to
sha1:sha1-base64. It also promotes cl-base64 as a direct dependency,
since it's already directly referenced in the source code.