Close #39: Replace ironclad dependency with sha1
Ironclad is a meaty dependency. This system takes the longest time on my
computer to compile my web application, and I've read comments online
that it adds ~19MB RAM usage in a Lisp image. Hunchensocket is
currently using ironclad to produce SHA1 digests. A single-purpose
library can fill this need instead.

This change switches the dependency from IRONCLAD:DIGEST-SEQUENCE to
SHA1:SHA1-BASE64. It also promotes cl-base64 as a direct dependency,
since it's already directly referenced in the source code.

* hunchensocket.asd: depend on base64 and sha1

* hunchensocket.lisp: Use SHA1:SHA1-BASE64 instead of 
  IRONCLAD:DIGEST-SEQUENCE

* package.lisp: No longer import IRONCLAD:DIGEST-SEQUENCE