Close #42: Fix binary mode
A nonexistent variable (`frame`) was accessed while writing to the
temporary file.  The mechanism has been updated to not use a temporary
file, but instead pass the whole binary message to the handler.

When signaling an error, a nonexistent slot (`status`) was accessed,
obscuring the error message that was printed when an unexpected binary
message was received.

 * hunchentoot.lisp: Fix issues